Move seq and aux methods to utils and import it on model_mommy init
