+++ model_mommy/recipe.py
@@ -3,26 +3,15 @@ from functools import wraps
 import inspect
 import itertools
 from . import mommy
-from .timezone import tz_aware
 from .exceptions import RecipeNotFound, RecipeIteratorEmpty
 
+# Enable seq to be imported from recipes
+from .utils import seq
+
 from six import string_types
-import datetime
-
-
-# Python 2.6.x compatibility code
-itertools_count = itertools.count
-try:
-    itertools_count(0, 1)
-except TypeError:
-    def count(start=0, step=1):
-        n = start
-        while True:
-            yield n
-            n += step
-    itertools_count =  count
 
 finder = mommy.ModelFinder()
+
 
 class Recipe(object):
     def __init__(self, _model, **attrs):
@@ -101,43 +90,6 @@ def foreign_key(recipe):
       will not be created during the recipe definition.
     """
     return RecipeForeignKey(recipe)
-
-
-def _total_secs(td):
-    """
-    python 2.6 compatible timedelta total seconds calculation
-    backport from
-    https://docs.python.org/2.7/library/datetime.html#datetime.timedelta.total_seconds
-    """
-    if hasattr(td, 'total_seconds'):
-        return td.total_seconds()
-    else:
-        #py26
-        return (td.microseconds + (td.seconds + td.days * 24 * 3600) * 10**6) / 10.0**6
-
-
-def seq(value, increment_by=1):
-    if type(value) in [datetime.datetime, datetime.date,  datetime.time]:
-        if type(value) is datetime.date:
-            date = datetime.datetime.combine(value, datetime.datetime.now().time())
-        elif type(value) is datetime.time:
-            date = datetime.datetime.combine(datetime.date.today(), value)
-        else:
-            date = value
-        # convert to epoch time
-        start = _total_secs((date - datetime.datetime(1970, 1, 1)))
-        increment_by = _total_secs(increment_by)
-        for n in itertools_count(increment_by, increment_by):
-            series_date = tz_aware(datetime.datetime.utcfromtimestamp(start + n))
-            if type(value) is datetime.time:
-                yield series_date.time()
-            elif type(value) is datetime.date:
-                yield series_date.date()
-            else:
-                yield series_date
-    else:
-        for n in itertools_count(increment_by, increment_by):
-            yield value + type(value)(n)
 
 
 class related(object):
